% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_ASE.R
\name{compare_ASE}
\alias{compare_ASE}
\title{compare_ASE}
\usage{
compare_ASE(res_edgeR_1, name1, res_edgeR_2, name2, output_intersection = TRUE)
}
\arguments{
\item{res_edgeR_1}{res_edgeR tibble of first comparison}

\item{name1}{character, name assigned to first comparison}

\item{res_edgeR_2}{res_edgeR tibble of second comparison}

\item{name2}{character, name assigned to second comparison}
}
\value{
a ggplot
}
\description{
compare_ASE
}
