% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plot_for_list.R
\name{coverage_plot_for_list}
\alias{coverage_plot_for_list}
\title{coverage_plot_for_list}
\usage{
coverage_plot_for_list(event_list, res_edgeR, se_object, control, treatment)
}
\arguments{
\item{event_list}{vector, of events to be plotted}

\item{res_edgeR}{output of get_DE_ASE}

\item{se_object}{output of get_ASE}

\item{control}{vector, name of control sample}

\item{treatment}{vector, name of treatment sample}
}
\value{
a coverage plot
}
\description{
coverage_plot_for_list
}
\examples{

event_list=c("CNAG_00332/AFR92466_Intron8/clean","CNAG_01940/AFR98135_Intron3/clean", "CNAG_02944/AFR93748_Intron1/clean" )
coverage_plot_for_list(event_list = event_list, res_edgeR = res_edgeR_bs181_5h, se_object = se_bs181_5h, control = "WT_5h",treatment = "WT_BS181_5h")

}
