% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_plots.R
\name{coverage_plots}
\alias{coverage_plots}
\title{coverage_plots}
\usage{
coverage_plots(res_edgeR, se_object, n_events = 3, control, treatment)
}
\arguments{
\item{res_edgeR}{output of get_DE_ASE}

\item{se_object}{output of get_ASE}

\item{n_events}{number of significant (FDR <0.05, deltaPSI>=0.05) events to be plotted}
}
\value{
a ggplot of coverage of ASE
}
\description{
coverage_plots
}
