% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ASE.R
\name{get_ASE}
\alias{get_ASE}
\title{get_ASE}
\usage{
get_ASE(reference_path, processed_bam_path, control, treatment, replicates = 3)
}
\arguments{
\item{reference_path}{character, path to the ensembl reference folder}

\item{processed_bam_path}{character, path to the output of the bam processed by spliceWiz}

\item{control}{character, name of the control sample}

\item{treatment}{character, name of treatment sample}

\item{replicates}{numeric, number of replicates}
}
\value{
An object of SummarisedExperiment
}
\description{
get_ASE
}
