% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_DE_ASE.R
\name{get_DE_ASE}
\alias{get_DE_ASE}
\title{get_DE_ASE}
\usage{
get_DE_ASE(
  se_filtered_object,
  control,
  treatment,
  PSI_cutoff = 0.05,
  LogFC_cutoff = 1,
  FDR_cutoff = 0.05,
  xlim = 6.5,
  plot_volcano = TRUE
)
}
\arguments{
\item{se_filtered_object}{summarised experiment object obtained from get_ASE function}

\item{control}{character, name of the control sample}

\item{treatment}{character, name of treatment sample}

\item{PSI_cutoff}{double, cutoff for delta_PSI, default: 0.05}

\item{LogFC_cutoff}{double, cutoff for Log2 fold-change, default: 1}

\item{FDR_cutoff}{double, cutoff on False Discovery Rate(FDR), default: 0.05}

\item{xlim}{double, X-axis limit for volcano plot}

\item{plot_volcano}{logical, to print volcano plot}
}
\value{
edgeR output
}
\description{
get_DE_ASE
}
