% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absPSI_ditribution.R
\name{plot_absPSI_ditribution}
\alias{plot_absPSI_ditribution}
\title{plot_absPSI_ditribution}
\usage{
plot_absPSI_ditribution(res_edgeR, plot_type = "boxplot")
}
\arguments{
\item{res_edgeR}{res_edgeR object}

\item{plot_type}{character, "densityplot" or "boxplot"}
}
\value{
a ggplot
}
\description{
plot_absPSI_ditribution
}
