% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano- Volcano plot for Alternate splicing Events}
\usage{
plot_volcano(res_edgeR)
}
\arguments{
\item{res_edgeR}{a res_edgeR object, out of get_DE_ASE}
}
\value{
two ggplots displaying DEG 1. for all the events 2. Classified according to type
}
\description{
plot_volcano- Volcano plot for Alternate splicing Events
}
\examples{

plot_volcano(res_edgeR_LP)

}
