% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicing_statistics.R
\name{splicing_statistics}
\alias{splicing_statistics}
\title{splicing_statistics}
\usage{
splicing_statistics(se_filtered_object, res_edgeR)
}
\arguments{
\item{se_filtered_object}{summarised experiment object obtained from get_ASE function}

\item{res_edgeR}{output of get_DE_ASE}
}
\value{
a barplot
}
\description{
splicing_statistics
}
